# service/Market/docs/swagger.yaml
basePath: /
definitions:
  github_com_Zifeldev_marketback_service_Market_internal_models.AddToCartRequest:
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
      size:
        type: string
    required:
    - product_id
    - quantity
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.CartItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      size:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.CartItemWithDetails:
    properties:
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      product_image:
        type: string
      product_price:
        type: number
      product_title:
        type: string
      quantity:
        type: integer
      size:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.CreateOrderRequest:
    properties:
      delivery_address:
        type: string
      payment_method:
        type: string
    required:
    - delivery_address
    - payment_method
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.Order:
    properties:
      created_at:
        type: string
      delivery_address:
        type: string
      id:
        type: integer
      payment_method:
        type: string
      payment_status:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.OrderItem:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      size:
        type: string
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.OrderWithItems:
    properties:
      created_at:
        type: string
      delivery_address:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.OrderItem'
        type: array
      payment_method:
        type: string
      payment_status:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.PaginationMeta'
    type: object
  github_com_Zifeldev_marketback_service_Market_internal_models.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.ProductWithDetails:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      price:
        type: number
      seller_id:
        type: integer
      seller_name:
        type: string
      sizes:
        items:
          type: string
        type: array
      status:
        type: string
      stock:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Zifeldev_marketback_service_Market_internal_models.UpdateCartItemRequest:
    properties:
      quantity:
        type: integer
      size:
        type: string
    required:
    - quantity
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Marketplace API with products, categories, cart, orders, seller and
    admin management
  title: Market Service API
  version: "1.0"
paths:
...
  /api/cart:
    get:
      consumes:
      - application/json
      description: Get current user's cart items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.CartItemWithDetails'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user cart
      tags:
      - cart
  /api/cart/items:
    post:
      consumes:
      - application/json
      description: Add a product to user's cart
      parameters:
      - description: Cart item data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.AddToCartRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.CartItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add item to cart
      tags:
      - cart
  /api/cart/items/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a cart item
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove item from cart
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: Update quantity of a cart item
      parameters:
      - description: Cart item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.UpdateCartItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.CartItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update cart item
      tags:
      - cart
  /api/categories:
    get:
      consumes:
      - application/json
      description: Get list of all product categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all categories
      tags:
      - categories
  /api/categories/{id}:
    get:
      consumes:
      - application/json
      description: Get category details by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get category by ID
      tags:
      - categories
  /api/products:
    get:
      consumes:
      - application/json
      description: Get paginated list of products with optional filters
      parameters:
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Filter by seller ID
        in: query
        name: seller_id
        type: integer
      - description: Filter by status
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.PaginatedResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all products
      tags:
      - products
  /api/products/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed product information
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.ProductWithDetails'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get product by ID
      tags:
      - products
...
  /api/user/orders:
    get:
      consumes:
      - application/json
      description: Get all orders for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.Order'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user orders
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: Create a new order from cart items
      parameters:
      - description: Order data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.OrderWithItems'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create order
      tags:
      - orders
  /api/user/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed order information
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Zifeldev_marketback_service_Market_internal_models.OrderWithItems'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"